Don't assume the JSON file can be installed in the installation folder
On Linux and macOS, the native manifest JSON file needs to be installed
in specific directories defined by the browser, which cannot be our
installation directory. While it's possible to generate the file in the
installation directory and to then move it into position, not having to
do an extra step helps.

Modify the -setup option to take a second argument, the directory in
which to install the JSON file -- but make it optional and default to
install folder as per the current behavior, so as to not break stuff.