Implement null verification of signature
We only really care that the signature is complete. In RSA, the
signature has the same length as the key; so therefore, verifying that
the length of the key is the same as the length of the signature should
suffice.

This allows us to drop the entire OpenSSL dependency, although we keep
it in for now in case we change our mind later.